Fix script not working for creating of PR
For the non-slack version, the "msg=" line was at the wrong point,
in the wrong if clause. Must somehow be moved. I also adapted the
syntax for the creation of the reviewer-list to seperate it by
comma and also put a colon after the reviewers to separate them
from the author.